fix(server): usar BigInteger nas colunas de timestamp em millis
Colunas *_millis (created/updated/started/ended) estavam como Integer
(32-bit). Timestamps em milissegundos estouram esse limite no Postgres;
o SQLite não aplica o tipo, então o bug só apareceu ao testar contra o
Supabase real (erro: NumericValueOutOfRange).

+++ server/src/models/activity.py
@@ -1,6 +1,6 @@
 import time
 
-from sqlalchemy import JSON, Boolean, Integer, String
+from sqlalchemy import JSON, BigInteger, Boolean, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from src.database import Base
@@ -18,10 +18,10 @@ class DBStudyActivity(Base):
     duration_minutes: Mapped[int] = mapped_column(Integer)
     duration_seconds: Mapped[int] = mapped_column(Integer)
     image_url: Mapped[str] = mapped_column(String)
-    started_at_millis: Mapped[int] = mapped_column(Integer)
-    ended_at_millis: Mapped[int] = mapped_column(Integer)
-    created_at_millis: Mapped[int] = mapped_column(Integer)
-    updated_at_millis: Mapped[int] = mapped_column(Integer, default=current_time_millis)
+    started_at_millis: Mapped[int] = mapped_column(BigInteger)
+    ended_at_millis: Mapped[int] = mapped_column(BigInteger)
+    created_at_millis: Mapped[int] = mapped_column(BigInteger)
+    updated_at_millis: Mapped[int] = mapped_column(BigInteger, default=current_time_millis)
     is_manual: Mapped[bool] = mapped_column(Boolean, default=False)
     author_id: Mapped[str] = mapped_column(String)
     author_name: Mapped[str] = mapped_column(String)
